[CI] Add GitHub Action checks for cache creator
Add a workflow that builds the ICD and the cache-creator to make
sure they build with both gcc and clang, with and without assertions.
This workflow uses the nightly docker images created by the LLPC cron
job:
https://github.com/GPUOpen-Drivers/llpc/actions?query=workflow%3A%22Build+AMDVLK+for+LLPC%22
to speed up the builds.

Add a workflow that ensures cache creator sources are properly
formatted.

This is mostly based on the public CI for LLPC:
https://github.com/GPUOpen-Drivers/llpc/actions.